post mejorado y funcionando

+++ app/Http/Controllers/PermissionController.php
@@ -15,7 +15,7 @@ class PermissionController extends Controller
      */
     public function index()
     {
-        // abort_if(Gate::denies('permission_index'), 403);
+        abort_if(Gate::denies('permissions.index'), 403);
         $permissions = Permission::paginate(10);
 
         return view('intermediary.permissions.index', compact('permissions'));
@@ -28,7 +28,7 @@ class PermissionController extends Controller
      */
     public function create()
     {
-        // abort_if(Gate::denies('permission_create'), 403);
+        abort_if(Gate::denies('permissions.create'), 403);
         return view('intermediary.permissions.create');
     }
 
@@ -53,7 +53,7 @@ class PermissionController extends Controller
      */
     public function show(Permission $permission)
     {
-        // abort_if(Gate::denies('permission_show'), 403);
+        abort_if(Gate::denies('permissions.show'), 403);
         return view('intermediary.permissions.show', compact('permission'));
     }
 
@@ -65,7 +65,7 @@ class PermissionController extends Controller
      */
     public function edit(Permission $permission)
     {
-        // abort_if(Gate::denies('permission_edit'), 403);
+        abort_if(Gate::denies('permissions.edit'), 403);
         return view('intermediary.permissions.edit', compact('permission'));
     }
 
@@ -91,7 +91,7 @@ class PermissionController extends Controller
      */
     public function destroy(Permission $permission)
     {
-        // abort_if(Gate::denies('permission_delete'), 403);
+        abort_if(Gate::denies('permissions.delete'), 403);
         
         $permission->delete();
         return redirect()->route('permissions.index');
