+++ app/Http/Controllers/RoleController.php
@@ -16,7 +16,7 @@ class RoleController extends Controller
      */
     public function index()
     {
-        // abort_if(Gate::denies('role_index'), 403);
+        abort_if(Gate::denies('roles.index'), 403);
         $roles = Role::paginate(10);
 
         return view('intermediary/roles/index', compact('roles'));
@@ -29,9 +29,9 @@ class RoleController extends Controller
      */
     public function create()
     {
-        // abort_if(Gate::denies('role_create'), 403);
+        abort_if(Gate::denies('roles.create'), 403);
         $permissions = Permission::all()->pluck('name', 'id');
-        // dd($permissions);
+        
         return view('intermediary/roles/create', compact('permissions'));
     }
 
@@ -44,10 +44,7 @@ class RoleController extends Controller
     public function store(Request $request)
     {
         $role = Role::create($request->only('name'));
-
-        // $role->permissions()->sync($request->input('permissions', []));
         $role->syncPermissions($request->input('permissions', []));
-
         
         return redirect()->route('roles.index');
     }
@@ -60,8 +57,9 @@ class RoleController extends Controller
      */
     public function show(Role $role)
     {
-        // abort_if(Gate::denies('role_show'), 403);
+        abort_if(Gate::denies('roles.show'), 403);
         $role->load('permissions');
+        
         return view('intermediary/roles/show', compact('role'));
     }
 
@@ -73,10 +71,10 @@ class RoleController extends Controller
      */
     public function edit(Role $role)
     {
-        // abort_if(Gate::denies('role_edit'), 403);
+        abort_if(Gate::denies('roles.edit'), 403);
         $permissions = Permission::all()->pluck('name', 'id');
         $role->load('permissions');
-        // dd($role);
+
         return view('intermediary/roles/edit', compact('role', 'permissions'));
     }
 
@@ -90,8 +88,6 @@ class RoleController extends Controller
     public function update(Request $request, Role $role)
     {
         $role->update($request->only('name'));
-
-        // $role->permissions()->sync($request->input('permissions', []));
         $role->syncPermissions($request->input('permissions', []));
 
         return redirect()->route('roles.index');
@@ -105,9 +101,9 @@ class RoleController extends Controller
      */
     public function destroy(Role $role)
     {
-        // abort_if(Gate::denies('role_delete'), 403);
-        
+        abort_if(Gate::denies('roles.delete'), 403);       
         $role->delete();
+
         return redirect()->route('roles.index');
     }
 }
